requirements.txt
blinker==1.6.2
certifi==2023.7.22
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pyparsing==3.0.9
requests==2.31.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
urllib3==2.0.2
Werkzeug==2.3.4
WTForms==3.0.1
bidict==0.22.1
colorama==0.4.6
dnspython==2.5.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
flask-starter==0.8.2
greenlet==3.0.3
h11==0.14.0
pymongo==4.6.1
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
simple-websocket==1.0.0
uuid==1.30
wsproto==1.2.0
gunicorn==20.1.0
Flask-CORS==4.0.0


